# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendus REST API",
        "version": "0.1"
    },
    "paths": {
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get list of customers",
                "description": "Returns a paginated list of customers with optional filtering based on various parameters.",
                "operationId": "getCustomers",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by customer name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filter by exact customer email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "partner_id",
                        "in": "query",
                        "description": "Filter by exact partner ID associated with the customer",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "discount",
                        "in": "query",
                        "description": "Filter by discount rate",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "Full name of the customer",
                                                        "type": "string",
                                                        "example": "XPTO"
                                                    },
                                                    "email": {
                                                        "description": "Email address of the customer",
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "xpto@email.com"
                                                    },
                                                    "partner_id": {
                                                        "description": "Identifier for the partner associated with the customer",
                                                        "type": "integer",
                                                        "format": "int64",
                                                        "example": 1
                                                    },
                                                    "discount": {
                                                        "description": "Discount rate applicable to the customer",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10.2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/customers?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/customers?page=10"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/customers?page=1",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/customers?page=3",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 16
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/customers"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "description": "Creates a new customer record in the database.",
                "operationId": "addCustomer",
                "requestBody": {
                    "description": "Customer data to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "Full name of the customer",
                                                        "type": "string",
                                                        "example": "XPTO"
                                                    },
                                                    "email": {
                                                        "description": "Email address of the customer",
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "xpto@email.com"
                                                    },
                                                    "partner_id": {
                                                        "description": "Identifier for the partner associated with the customer",
                                                        "type": "integer",
                                                        "format": "int64",
                                                        "example": 1
                                                    },
                                                    "discount": {
                                                        "description": "Discount rate applicable to the customer",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10.2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error messsage"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "Object with all the errors by field"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Customer": {
                "title": "Customer",
                "description": "Schema for Customer data",
                "required": [
                    "name",
                    "vat",
                    "email",
                    "address",
                    "partner_id",
                    "discount"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the customer",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Full name of the customer",
                        "type": "string"
                    },
                    "vat": {
                        "description": "VAT number of the customer",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the customer",
                        "type": "string",
                        "format": "email"
                    },
                    "address": {
                        "description": "Physical address of the customer",
                        "type": "string"
                    },
                    "partner_id": {
                        "description": "Identifier for the partner associated with the customer",
                        "type": "integer",
                        "format": "int64"
                    },
                    "discount": {
                        "description": "Discount rate applicable to the customer",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
        }
    }
}